Remove unnecessary component layer for thresholds overview side panel
